refactor: Added tests for blueprints and test-support
